Re-do the left-hand side menu to allow for two levels
Rather than `menu_pages` being just a list/queryset of pages it's
now a structure, with each item like `{'page': Page, 'children': []}`

Each element in `children` can be a Page. We're assuming that there
can't be more than two levels, given that's what's shown on the designs.

For #42
For #32
For #20
For #15
For #29