Make it so JobsIndexPage has to live within SectionListingPage ("About")
Previously it could only be a child of the HomePage because in the designs
it's not "in" a section. But this means that when clicking a tag on a
job description, there's nowhere to go - the theme listing / taxonomy
page shows content within a section.

So we've moved the Careers page to be within About, so that its child
content (jobs) can appear within that section's themed listing page.