Add get_subnav_top_level_page() template extension to fix left-hand menu
We were relying on the concept of `section_page` to be able to tell
where the current page was within the navbar hierarchy. But (a) there
are now pages with menus that can exist outside of Section pages and
(b) the navbar hierarchy no longer matches the Wagtail Pages
hierarchy, which confuses things further.

So, before rendering the left-hand menu we need to find out which of
the top-level menu navbar menu items, if any, the curent page is beneath.
Then we know which of those parts of the navbar to display as the
left-hand menu.

So the `get_subnav_top_level_page()` template extension will cycle
through the navbar pages and return the Page that is the top-most
one above the current page, if any.

(Felt a bit like this should/could be done somewhere in the Nav settings
code, but that only returns a single navbar data structure and messing
with that seems like more trouble than it's worth. This gets us the
required info in a separate way without messing with anything else.)

For #83