Add card_blurb property to BasePage to display role for TeamProfilePage cards
When displaying cards for TeamProfilePages we want to display `role` in
place of `blurb`. Rather than put a special case into the card template,
I've added a new `card_blurb` `@cached_property` on `BasePage` which,
by default, returns the `blurb`.

On `TeamProfilePage` it returns the `role`.

We then use `card_blurb` in the card template instead of the plain
`blurb`.

For #84