Fix bug in left-hand menu
If a Page (as opposed to a View) displayed a left-hand menu, but didn't
exist anywhere in the top navigation, it would end up with a left-hand
menu featuring whatever the last item was in the navbar (whether it
was a single Link Page, or the Page at the top of a Mega Nav). Which
might, on occasion, be convenient. But it was unpredictable and not
planned.