Fix top-level pages not being "on" in topnav
When visiting one of the top-level pages that had a meganav (like
Research, Implement, Impact), the item in the top navigation wasn't
being marked as "on". It wasn't getting the `aria-current="page"`
attribute.

For some reason, when doing this in `navigation.jinja`:

    obj.page == page.section_page

where `obj` is the navbar object representing that top level link,
and `page` is the currently-viewed page, it was returning `False`, when
it should be `True` for the current page.

I've no idea why. Both were the same page. Both have the same `pk`.

So I've changed it to:

	obj.page.pk == page.section_page.pk

which works. I *assume* something about the Page object changed somewhere
so that however the objects are compared, means they were no longer
considered equal.