Remove 'section_page' from get_url() in tags/categories
We were no longer user the `section_page` passed in to the `get_url()`
method in tags and categories to generate their URLs - at some point
it must have been decided that these were no longer within sections.
So it simplifies things if we just remove that from that from the method
and all the places where the method is called.