Improve highlighting of top-level item in top nav menu
We were using `BasePage.section_page` property in `navigation.jinja`
to determine which of the top-level links in the topnav should be
highlighted. But `BasePage.section_page` works off the Wagtail
Page hierarchy, which is no longer the same as the navigation
hierarchy, so this wasn't working.

Instead we can now use the same function used to get the left-hand menu,
now named `get_top_level_navpage()` which uses the navbar settings
hierarchy to return the top-most Page above the current page.

Added a tweak for Publication front and inner pages, because they are
never within the navigation hierarchy, so that they use their parent/
grandparent (the PublicationIndexPage) to determine their position
in the nav hierarchy instead.

For #142