Add a second set of translation files to override words Django insists on translating
There are some words that we want to keep in English on other
languages' sites. e.g. "Home" and "Search" in the navbar.

Even if setting the translations for these in the languages' `.po`
files like:

    msgid "Search"
	msgstr "Search"

The terms often get translated into the local language.

But, for some reason, adding a second set of translation files
(`app/local_extra/`) and adding these non-translations into those,
forces the words into English.

From https://stackoverflow.com/a/41945558/250962

For #360